backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.5.3
psycopg[binary]
redis==5.0.1
python-dotenv==1.0.0
werkzeug==3.0.1